# migration.openapi.yaml
openapi: 3.0.0
info:
  title: ziplinee-ci-api
  version: 0.0.0
  description: Migration API
paths:
  /api/auth/client/login:
    post:
      operationId: authentication
      description: Authenticate with client id and secret
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Forbidden
        '403':
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequest:
      type: object
      required:
      - clientID
      - clientSecret
      additionalProperties: false
      properties:
        clientID:
          type: string
        clientSecret:
          type: string
    AuthResponse:
      type: object
      required:
      - token
      additionalProperties: true
      properties:
        token:
          type: string
...